Address comments and simplify test runner
